+++ packages/Flow/src/package-test.ts
@@ -26,6 +26,7 @@ import './tests/rerun-node-tests';
 import './tests/func-editor-tests';
 import './tests/entity-tests';
 import './tests/editor-bridge-tests';
+import './tests/invalidation-tests';
 
 export let _package = new DG.Package();
 export { tests };